harden trimmomatic to unpaired reads, remove entry point, move code to Tool, move test to Tool
